Tamobla Game
Open Tambola.HTML file in visual studio code and run with live server to play the game Tambola.Html file contains the main HTML content of the game, which contains Game board, ticket,reset and next buttons. once the ticket button is pressed it will head over to ticket.html file which is created to generate tickets. How to play the game. Click on ticket button to generate tickets and Press next button to start a random number on the main board, once a random number is highlighted on the main board, check if the number matches with the number existing in ticket page, if the number matches click on the matched number to block it..like wise generate the randum numbers until all your numbers in ticket is filled. once the ticket is filled you will get a game over message.